refactor: 사용자 체크리스트 비즈니스 로직 커스텀훅으로 분리
